Reorganized contracts
