--- contracts/Bindings.sol
@@ -1,11 +1,0 @@
-pragma solidity ^0.4.25;
-
-import "./DarknodePayment.sol";
-import "./DarknodePaymentStore.sol";
-import "./DarknodeRegistry.sol";
-import "./RepublicToken.sol";
-
-/// @notice Bindings imports all of the contracts for generating bindings.
-/* solium-disable no-empty-blocks */
-contract Bindings {
-}
