+++ contracts/LinkedList.sol
