+++ contracts/test/Bindings.sol
@@ -1,0 +1,11 @@
+pragma solidity ^0.4.25;
+
+import "../DarknodePayment/DarknodePayment.sol";
+import "../DarknodePayment/DarknodePaymentStore.sol";
+import "../DarknodeRegistry/DarknodeRegistry.sol";
+import "../RepublicToken.sol";
+
+/// @notice Bindings imports all of the contracts for generating bindings.
+/* solium-disable no-empty-blocks */
+contract Bindings {
+}
