+++ generateGoBindings.sh
@@ -8,7 +8,7 @@ mkdir ./contracts/openzeppelin-solidity
 cp -r ./node_modules/openzeppelin-solidity/contracts ./contracts/openzeppelin-solidity/contracts
 
 ### GENERATE BINDINGS HERE ###
-abigen --sol ./contracts/Bindings.sol -pkg bindings --out bindings.go
+abigen --sol ./contracts/test/Bindings.sol -pkg bindings --out bindings.go
 
 # Revert setup
 sed -i.bak -e 's/".\/openzeppelin-solidity\/contracts\//"openzeppelin-solidity\/contracts\//' contracts/*.sol
